requirements.txt
Django==3.2.9
msgspec==0.3.2
pyzmq==22.3.0
tornado==6.1
